Repository and Models Finished

+++ Publicaciones/Publicaciones.Domain/Entities/Jobs.cs
@@ -9,7 +9,7 @@ namespace Publicaciones.Domain.Entities
 {
     public class Jobs : BaseEntity
     {
-        public short JobID { get; set; }
+        public int JobID { get; set; }
         public string JobDescription { get; set; }
         public byte Minlvl { get; set; }
         public byte Maxlvl { get; set; }
